move plotter from /io to /analyzer. Tested by using real 2019 data-tickets generating the final report.
